Issa xss fix 2 (#149)
This covers a case that should have been covered in #148 but was
missed. It was included in the grep search results (see PR description
for the commands used) but got missed in the manual filtering step
last time.

Also, fix SQL syntax error; MySQL doesn't get tripped by this (it
successfully converts double hyphen to single hyphen when parsing) but
other forms of SQL such as MariaDB do get tripped.